chore: split tests
Unfortunately the winit event loop doesn't fully exit making it impossible to run multiple tests inside the same test file.
